requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
requests>=2.31.0
joblib>=1.3.0
PyYAML>=6.0.1
streamlit>=1.42.0
shap>=0.47.0
matplotlib>=3.8.0
pytest>=8.0.0
